backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
defusedxml==0.7.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
httpie==3.2.4
idna==3.10
inflection==0.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.6.3
packaging==25.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
PySocks==1.7.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==80.9.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
